notebooks/autoencoder/requirements_evaluation.txt
# Optional dependencies for comprehensive autoencoder evaluation
# Install with: pip install -r requirements_evaluation.txt

# For advanced metrics
torchmetrics>=0.11.0
lpips>=0.1.4

# For statistical analysis
scipy>=1.9.0
scikit-learn>=1.1.0

# For visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Core dependencies (if not already installed)
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
tqdm>=4.64.0

# For dataset handling
Pillow>=9.2.0 